Fix checkbox bug
